crispr_backend/requirements.txt
biopython==1.83
fastapi>=0.68.0
uvicorn>=0.15.0
